requirements.txt
fastapi
uvicorn
pydantic
requests
python-dotenv
sqlite3
marker-pdf
pytesseract
pillow
